+++ packages/style/src/fonts/fonts.css.ts
@@ -1,5 +1,5 @@
 import { globalFontFace } from "@vanilla-extract/css";
-import { interFontFaces } from "./inter/fontFaces";
+import { interFontFaces } from "./inter/font-faces";
 import { interFontFamily } from "./fonts";
 
 // We register fonts using globalFontFace so that @mp/style can be imported on the server without having vanilla-extract
